フィード生成機能を独立化
